Adding author and authentication and authorization

+++ routes/campground.js
@@ -4,6 +4,7 @@ const catchAsync=require("../utils/catchAsync");
 const ExpressError=require("../utils/ExpressError");
 const Campground=require('../models/campground');
 const {campgroundSchema}=require('../schemas');
+const { isLoggedIn } = require('../middleware');
 
 const validateCampground=(req,res,next)=>{
     const {error}=campgroundSchema.validate(req.body);
@@ -20,19 +21,20 @@ router.get('/',catchAsync(async(req,res)=>{
     res.render('campgrounds/index',{campgrounds})
  }))
  
- router.get('/new',(req,res)=>{
+ router.get('/new',isLoggedIn,(req,res)=>{
      res.render('campgrounds/new');  
  })
  
- router.post('/',validateCampground,catchAsync(async (req,res,next)=>{
+ router.post('/',isLoggedIn,validateCampground,catchAsync(async (req,res,next)=>{
      const campground=new Campground(req.body.campground);
+     campground.author=req.user._id;
      await campground.save();
-     req.flash('success','Campgrund added Successfully!');
+     req.flash('success','Campground added Successfully!');
      res.redirect(`/campgrounds/${campground._id}`)
  }))
  
  router.get('/:id',catchAsync(async(req,res)=>{
-     const campground=await Campground.findById(req.params.id).populate('reviews');
+     const campground=await Campground.findById(req.params.id).populate('reviews').populate('author');
      if(!campground)
      {
         req.flash('error','Cannot find the campground!');
@@ -41,27 +43,45 @@ router.get('/',catchAsync(async(req,res)=>{
      res.render('campgrounds/show',{campground});
  }))
  
- router.get('/:id/edit',catchAsync(async(req,res)=>{
-     const campground=await Campground.findById(req.params.id);
+ router.get('/:id/edit',isLoggedIn,catchAsync(async(req,res)=>{
+    const {id}=req.params; 
+    const campground=await Campground.findById(id);
      if(!campground)
      {
         req.flash('error','Cannot find the campground!');
         return res.redirect('/campgrounds');
      }
+     if(!campground.author.equals(req.user._id))
+     {
+        req.flash('error', "You don't have the Permission to do that");
+        return res.redirect(`/campgrounds/${campground._id}`)
+     }
      res.render('campgrounds/edit',{campground});
  }))
  
- router.put('/:id',validateCampground,catchAsync(async(req,res)=>{
+ router.put('/:id',isLoggedIn,validateCampground,catchAsync(async(req,res)=>{
      const {id}=req.params;
-     const campground= await Campground.findByIdAndUpdate(id,{...req.body.campground});
-     req.flash('success','Successfull Updated!');
+     const campground=await Campground.findById(id);
+     if(!campground.author.equals(req.user._id))
+     {
+        req.flash('error', "You don't have the Permission to do that");
+        return res.redirect(`/campgrounds/${campground._id}`)
+     }
+     const camp= await Campground.findByIdAndUpdate(id,{...req.body.campground});
+     req.flash('success','Successfully Updated!');
      res.redirect(`/campgrounds/${campground._id}`)
  }))
  
- router.delete('/:id',catchAsync(async(req,res)=>{
-     const {id}=req.params;
+ router.delete('/:id',isLoggedIn,catchAsync(async(req,res)=>{
+    const {id}=req.params;
+    const campground=await Campground.findById(id);
+    if(!campground.author.equals(req.user._id))
+    {
+       req.flash('error', "You don't have the Permission to do that");
+       return res.redirect(`/campgrounds/${campground._id}`)
+    }
      await Campground.findByIdAndDelete(id);
-     req.flash('success','Review Successfully Deleted!');
+     req.flash('success','Campground deleted Successfully!');
      res.redirect('/campgrounds');
  }))
